Refactor commands to a separate module
